Add startTLS support.
Change the options to `create-connection` to look for a `tls?` key, in
addition to `ssl?`. If `tls?` is truthy, then have the connection
perform the startTLS invocation after it's created. If both `tls?` and
`ssl?` are truthy, throw an IllegalArgumentException, as the two are
mutually exclusive.

Add a test to ensure that we can connect to the server and perform the
startTLS invocation without errors (this is not a strong test that
startTLS works, but it's as strong as the SSL tests, and a truly strong
test would be much more work).