Don't AOT compile.
We're a library, so we shouldn't be AOT compiling our code since there's
no benefit. Downstream applications should be able to decide whether
they want to AOT or not, and if they do our code will be transitively
compiled.

AOT'd code in jars only increases the size of jars, and opens up the
possibility for strange errors in downstream projects when they use
a newer version of clojure than the AOT'd code does.