Add case to client/create-modification to support directly supplying bytes
For some operations, such as setting unicodePwd for an Active Directory entry,
the simplest way to provide the value of an attribute is as a byte array due to
encoding issues. As-is, when `values` is a byte array, it matches `:else` and
is processed with `str`, which is substantively incorrect.

The current behavior is particularly confusing because the underlying LDAP SDK
supports providing a byte array out of the box, and because the value that will
be written is not even an incorrectly-encoded version of `values`; it will be a
string of the form "[B@<reference>".